- Adds Butterfly Knives skill card, sends out 3 projectiles when doing a "flap" slash - Killing an enemy with a dagger or a mace no longer causes one final attack - Adds animated helmet support to the player - Adds a new generic Grunt behavior that Grunts, Twitches and Bulwarks all share - Adds a new Executioner enemy that deals area damage while swinging the mace - Adds Blood Thorn skill card, sends out a blood thorn when performing a slash while blade is bloody. Blood thorns deal damage over time. - Adds Vampire Blade skill card, heals the player when sheathing a bloody blade. - Adds Bloodfire Blade skill card, sword ignites after being bloody, deals +1 damage on slash and instakills on stab.

+++ 7DRL_2021/Results/ActionTurn.cs
@@ -15,7 +15,7 @@ namespace _7DRL_2021.Results
         private Slider Frame;
         public float Slide => Frame.Slide;
 
-        public ActionTurn(ICurio origin, float turnAngle, int time)
+        public ActionTurn(ICurio origin, float turnAngle, float time)
         {
             Origin = origin;
             TurnAngle = turnAngle;
